+++ test/datagen.js
@@ -24,6 +24,9 @@ buster.testCase('datagen - generate', {
     this.mockFs = this.mock(fs);
   },
   'should send default message properties when none is specified': function (done) {
+    this.mockFs.expects('existsSync').once().withExactArgs('header').returns(true);
+    this.mockFs.expects('existsSync').once().withExactArgs('segment').returns(true);
+    this.mockFs.expects('existsSync').once().withExactArgs('footer').returns(true);
     this.mockFs.expects('readFileSync').once().withExactArgs('header').returns('header template');
     this.mockFs.expects('readFileSync').once().withExactArgs('segment').returns('segment template');
     this.mockFs.expects('readFileSync').once().withExactArgs('footer').returns('footer template');
@@ -47,6 +50,9 @@ buster.testCase('datagen - generate', {
     datagen.generate();
   },
   'should send specified message properties when they are provided': function (done) {
+    this.mockFs.expects('existsSync').once().withExactArgs('header').returns(true);
+    this.mockFs.expects('existsSync').once().withExactArgs('segment').returns(true);
+    this.mockFs.expects('existsSync').once().withExactArgs('footer').returns(true);
     this.mockFs.expects('readFileSync').once().withExactArgs('header').returns('header template');
     this.mockFs.expects('readFileSync').once().withExactArgs('segment').returns('segment template');
     this.mockFs.expects('readFileSync').once().withExactArgs('footer').returns('footer template');
@@ -70,5 +76,24 @@ buster.testCase('datagen - generate', {
 
     var datagen = new DataGen();
     datagen.generate({ genId: 'somegenid', numSegments: 3, numWorkers: 2, outFile: 'someoutfile' });
+  },
+  'should default to empty string when any of the template file does not exist': function (done) {
+    this.mockFs.expects('existsSync').once().withExactArgs('header').returns(false);
+    this.mockFs.expects('existsSync').once().withExactArgs('segment').returns(false);
+    this.mockFs.expects('existsSync').once().withExactArgs('footer').returns(false);  
+    this.stub(process, 'pid', 12345);
+
+    var mockWorker = {
+      send: function (opts) {
+        assert.equals(opts.templates.header, '');
+        assert.equals(opts.templates.segment, '');
+        assert.equals(opts.templates.footer, '');
+        done();
+      }
+    };
+    this.mockChildProcess.expects('fork').once().returns(mockWorker);
+
+    var datagen = new DataGen();
+    datagen.generate();
   }
 });