+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.7 (SNAPSHOT)
-*
+* Header and footer template files are now optional
 
 ### 0.0.6
 * Replace config command with init command